+++ examples/simple-vpc/main.tf
@@ -2,6 +2,11 @@ provider "aws" {
   region = "eu-west-1"
 }
 
+data "aws_security_group" "default" {
+  name   = "default"
+  vpc_id = "${module.vpc.vpc_id}"
+}
+
 module "vpc" {
   source = "../../"